requirements.txt
appdirs==1.4.0
blessings==1.6
bpython==0.16
bzr==2.7.0.dev1
BzrTools==2.6.0
chardet==2.0.1
colorama==0.2.5
configobj==4.7.2
cryptography==1.7.2
curtsies==0.2.11
decorator==3.4.0
Django==1.9
django-filter==1.0.1
djangorestframework==3.5.4
duplicity==0.6.23
enum34==1.1.6
greenlet==0.4.12
html5lib==0.999
httplib2==0.8
idna==2.2
ipaddress==1.0.18
ipython==1.2.1
keyring==3.5
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lockfile==0.8
lpthw.web==1.1
lxml==3.7.3
Markdown==2.6.8
matplotlib==1.3.1
mercurial==2.8.2
ndg-httpsclient==0.4.2
numpy==1.8.2
oauth==1.0.1
packaging==16.8
paramiko==1.10.1
pexpect==3.1
Pillow==2.3.0
pyasn1==0.2.2
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.12.0
pygpgme==0.3
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-apt===0.9.3.5ubuntu2
python-dateutil==1.5
pytz==2012rc0
requests==2.2.1
scipy==0.13.3
SecretStorage==2.0.0
simplegeneric==0.8.1
simplejson==3.3.1
six==1.10.0
stevedore==0.14.1
urllib3==1.7.1
virtualenv==1.11.4
virtualenv-clone==0.2.4
virtualenvwrapper==4.1.1
wadllib==1.3.2
wcwidth==0.1.7
zope.interface==4.0.5
